rename guest templates to tccp
