+++ service/controller/v24/key/key.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/giantswarm/aws-operator/service/controller/v24/templates/cloudconfig"
 	"github.com/giantswarm/aws-operator/service/controller/v24/templates/cloudformation/cpf"
-	"github.com/giantswarm/aws-operator/service/controller/v24/templates/cloudformation/guest"
 	"github.com/giantswarm/aws-operator/service/controller/v24/templates/cloudformation/hostpre"
+	"github.com/giantswarm/aws-operator/service/controller/v24/templates/cloudformation/tccp"
 )
 
 const (
@@ -152,21 +152,21 @@ func CloudConfigSmallTemplates() []string {
 
 func CloudFormationGuestTemplates() []string {
 	return []string{
-		guest.AutoScalingGroup,
-		guest.IAMPolicies,
-		guest.Instance,
-		guest.InternetGateway,
-		guest.LaunchConfiguration,
-		guest.LoadBalancers,
-		guest.Main,
-		guest.NatGateway,
-		guest.LifecycleHooks,
-		guest.Outputs,
-		guest.RecordSets,
-		guest.RouteTables,
-		guest.SecurityGroups,
-		guest.Subnets,
-		guest.VPC,
+		tccp.AutoScalingGroup,
+		tccp.IAMPolicies,
+		tccp.Instance,
+		tccp.InternetGateway,
+		tccp.LaunchConfiguration,
+		tccp.LoadBalancers,
+		tccp.Main,
+		tccp.NatGateway,
+		tccp.LifecycleHooks,
+		tccp.Outputs,
+		tccp.RecordSets,
+		tccp.RouteTables,
+		tccp.SecurityGroups,
+		tccp.Subnets,
+		tccp.VPC,
 	}
 }
