+++ service/controller/v24/templates/cloudformation/tccp/iam_policies.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const IAMPolicies = `{{define "iam_policies"}}
 {{- $v := .Guest.IAMPolicies }}
