+++ service/controller/v24/templates/cloudformation/tccp/instance.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const Instance = `{{ define "instance" }}
 {{- $v := .Guest.Instance }}
