+++ service/controller/v24/templates/cloudformation/tccp/launch_configuration.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const LaunchConfiguration = `{{define "launch_configuration"}}
 {{- $v := .Guest.LaunchConfiguration }}
