+++ service/controller/v24/templates/cloudformation/tccp/main.go
@@ -1,11 +1,11 @@
-package guest
+package tccp
 
 const Main = `{{define "main"}}AWSTemplateFormatVersion: 2010-09-09
 Description: Main Guest CloudFormation stack.
 Parameters:
   VersionBundleVersionParameter:
     Type: String
-    Description: Sets the VersionBundleVersion used to generate the template. 
+    Description: Sets the VersionBundleVersion used to generate the template.
 Resources:
   {{template "vpc" .}}
   {{template "iam_policies" .}}
