+++ service/controller/v24/templates/cloudformation/tccp/nat_gateway.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const NatGateway = `{{define "nat_gateway"}}
   {{- $v := .Guest.NATGateway }}
