+++ service/controller/v24/templates/cloudformation/tccp/outputs.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const Outputs = `{{define "outputs"}}
 {{- $v := .Guest.Outputs }}
