+++ service/controller/v24/templates/cloudformation/tccp/route_tables.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const RouteTables = `{{ define "route_tables" }}
 {{- $v := .Guest.RouteTables }}
