+++ service/controller/v24/templates/cloudformation/tccp/vpc.go
@@ -1,4 +1,4 @@
-package guest
+package tccp
 
 const VPC = `{{define "vpc"}}
 {{- $v := .Guest.VPC }}
